+++ docs/examples/02_description_models/files/02_robot_artist_blender.py
@@ -1,7 +1,3 @@
-import compas
-
-compas.PRECISION = "12f"
-
 from compas_robots import RobotModel
 from compas_robots.resources import GithubPackageMeshLoader
 from compas_robots.blender import RobotModelArtist
@@ -14,6 +10,6 @@ github = GithubPackageMeshLoader(r, p, b)
 urdf = github.load_urdf("irb6640.urdf")
 
 model = RobotModel.from_urdf_file(urdf)
-model.load_geometry(github)
+model.load_geometry(github, precision=12)
 
 RobotModelArtist(model).draw_visual()