+++ docs/examples/02_description_models/files/02_robot_artist_rhino.py
@@ -1,17 +1,15 @@
-import compas
-compas.PRECISION = '12f'
-
-from compas_robots import *
+from compas_robots import RobotModel
+from compas_robots.resources import GithubPackageMeshLoader
 from compas_rhino.artists import RobotModelArtist
 
-r = 'ros-industrial/abb'
-p = 'abb_irb6600_support'
-b = 'kinetic-devel'
+r = "ros-industrial/abb"
+p = "abb_irb6600_support"
+b = "kinetic-devel"
 
 github = GithubPackageMeshLoader(r, p, b)
-urdf = github.load_urdf('irb6640.urdf')
+urdf = github.load_urdf("irb6640.urdf")
 
 robot = RobotModel.from_urdf_file(urdf)
-robot.load_geometry(github)
+robot.load_geometry(github, precision=12)
 
 RobotModelArtist(robot).draw_visual()