+++ docs/examples/02_description_models/files/02_robot_from_github.py
@@ -1,9 +1,5 @@
-import compas
 from compas_robots.resources import GithubPackageMeshLoader
 from compas_robots import RobotModel
-
-# Set high precision to import meshes defined in meters
-compas.PRECISION = "12f"
 
 # Select Github repository, package and branch where the model is stored
 repository = "ros-industrial/abb"
@@ -17,6 +13,6 @@ urdf = github.load_urdf("irb6640.urdf")
 model = RobotModel.from_urdf_file(urdf)
 
 # Also load geometry
-model.load_geometry(github)
+model.load_geometry(github, precision=12)
 
 print(model)